+++ apps/web/src/components/logs/log-trace-timeline.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef } from "react"
 import { Result, useAtomValue } from "@/lib/effect-atom"
 import { Skeleton } from "@maple/ui/components/ui/skeleton"
 import { cn } from "@maple/ui/utils"
-import { getSeverityColor } from "@/lib/severity"
+import { getSeverityColor } from "@maple/ui/lib/severity"
 import type { Log, LogsResponse } from "@/api/warehouse/logs"
 import type { SpanHierarchyResponse } from "@/api/warehouse/traces"
 import { listLogsResultAtom, getSpanHierarchyResultAtom } from "@/lib/services/atoms/warehouse-query-atoms"
