+++ apps/web/src/components/logs/logs-table.tsx
@@ -8,7 +8,7 @@ import { LogDetailSheet } from "./log-detail-sheet"
 import type { LogsSearchParams } from "@/routes/logs"
 import { useTimezonePreference } from "@/hooks/use-timezone-preference"
 import { formatCompactTimeInTimezone } from "@/lib/timezone-format"
-import { getSeverityColor } from "@/lib/severity"
+import { getSeverityColor } from "@maple/ui/lib/severity"
 import { useInfiniteLogs, FETCH_THRESHOLD } from "@/hooks/use-infinite-logs"
 import { pickImportantAttributes } from "@/lib/log-attributes"
 import { LogAttributeChip } from "./log-attribute-chip"
