+++ apps/web/src/routes/traces/$traceId.tsx
@@ -7,7 +7,7 @@ import { TraceId } from "@maple/domain"
 import { DashboardLayout } from "@/components/layout/dashboard-layout"
 import { TraceReplayLink } from "@/components/replays/trace-replay-link"
 import { QueryErrorState } from "@/components/common/query-error-state"
-import { TraceViewTabs } from "@/components/traces/trace-view-tabs"
+import { TraceViewTabs } from "@maple/ui/components/traces/trace-view-tabs"
 import { SpanDetailPanel } from "@/components/traces/span-detail-panel"
 import { Badge } from "@maple/ui/components/ui/badge"
 import { Skeleton } from "@maple/ui/components/ui/skeleton"
@@ -16,8 +16,8 @@ import { formatDuration } from "@/lib/format"
 import { getServiceLegendColor } from "@maple/ui/lib/colors"
 import { type Span, type SpanNode } from "@/api/warehouse/traces"
 import { getSpanHierarchyResultAtom } from "@/lib/services/atoms/warehouse-query-atoms"
-import { findSpanById } from "@/components/traces/flow-utils"
-import { HttpSpanLabel } from "@/components/traces/http-span-label"
+import { findSpanById } from "@maple/ui/components/traces/flow-utils"
+import { HttpSpanLabel } from "@maple/ui/components/traces/http-span-label"
 import { TraceIdBadge } from "@/components/traces/trace-id-badge"
 import { CopyableBadge } from "@/components/common/copyable-badge"
 import { getHttpInfo } from "@maple/ui/lib/http"
