+++ packages/ui/src/components/icons/index.ts
@@ -14,12 +14,15 @@ export { CircleCheckIcon } from "./circle-check"
 export { CircleInfoIcon } from "./circle-info"
 export { CircleXmarkIcon } from "./circle-xmark"
 export { CodeIcon } from "./code"
+export { DatabaseIcon } from "./database"
 export { DotsIcon } from "./dots"
 export { EyeIcon } from "./eye"
+export { FireIcon } from "./fire"
 export { LayoutMoveToLeftIcon } from "./layout-move-to-left"
 export { LayoutMoveToRightIcon } from "./layout-move-to-right"
 export { LoaderIcon } from "./loader"
 export { MagnifierIcon } from "./magnifier"
+export { MenuIcon } from "./menu"
 export { MinusIcon } from "./minus"
 export { NetworkNodesIcon } from "./network-nodes"
 export { PulseIcon } from "./pulse"
