+++ packages/ui/src/components/icons/menu.tsx
@@ -1,0 +1,23 @@
+import type { IconProps } from "./icon"
+
+const paths: ReadonlyArray<string> = ["M4 6H20", "M4 12H20", "M4 18H20"]
+
+function MenuIcon({ size = 24, className, ...props }: IconProps) {
+	return (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			viewBox="0 0 24 24"
+			width={size}
+			height={size}
+			className={className}
+			fill="none"
+			aria-hidden="true"
+			{...props}
+		>
+			{paths.map((d, i) => (
+				<path key={i} d={d} stroke="currentColor" strokeWidth="2" strokeLinecap="square" />
+			))}
+		</svg>
+	)
+}
+export { MenuIcon }
