+++ packages/ui/src/components/traces/pixel-duration-bar.tsx
