+++ packages/ui/src/components/traces/trace-timeline-canvas.tsx
