+++ packages/ui/src/components/traces/trace-timeline.tsx
@@ -1,10 +1,10 @@
 import * as React from "react"
 import * as ReactDOM from "react-dom"
 
-import { ChevronExpandYIcon } from "@/components/icons"
-import { Button } from "@maple/ui/components/ui/button"
-import { getServiceLegendColor } from "@maple/ui/colors"
-import { useContainerSize } from "@maple/ui/hooks/use-container-size"
+import { ChevronExpandYIcon } from "../icons"
+import { Button } from "../ui/button"
+import { getServiceLegendColor } from "../../lib/colors"
+import { useContainerSize } from "../../hooks/use-container-size"
 import { useTraceView } from "./trace-view-context"
 import { useTraceTimeline } from "./use-trace-timeline"
 import { collectAllCollapsibleIds } from "./auto-collapse"
