+++ packages/ui/src/components/traces/use-canvas-crosshair.ts
