+++ packages/ui/src/components/traces/use-canvas-hit-test.ts
