+++ packages/ui/src/components/traces/use-canvas-setup.ts
